:recycle: [Refactor] 라우터 관련 함수 분리
